# ARROW-8996: [C++] Add AVX version for aggregate sum/mean with runtime dispatch