… cpp/src/arrow/CMakeLists.txt
… cpp/src/arrow/compute/kernels/aggregate_sum_avx2.cc
… cpp/src/arrow/compute/kernels/aggregate_sum_avx512.cc
… cpp/src/arrow/compute/kernels/aggregate_test.cc
… cpp/src/arrow/compute/registry.cc
… cpp/src/arrow/compute/registry_internal.h
… cpp/src/arrow/util/bit_block_counter.cc
… cpp/src/arrow/util/bit_block_counter.h
… cpp/src/arrow/util/cpu_info.cc
… cpp/src/arrow/util/cpu_info.h
… cpp/src/arrow/util/spaced.h