1. Add AVX2/AVX512 build version of aggregate sum/mean function.
     Use set_source_files_properties to append the SIMD build option. 
     Register the SIMD path at runtime by CPU feature.
     Move the sum/mean implementation to header file thus SIMD version can reuse.
     Also rework the implementation based on bit_block_counter.
2. Introduce ARROW_USER_SIMD_LEVEL env for test scope.
     ex "ARROW_USER_SIMD_LEVEL=sse4_2 make unittest" on a AVX2 capable device.