@jianxind 2020-07-08
@wesm @emkornfield @pitrou 

Ping, could you help to review this approach when you are available? Appreciate your help.