@jianxind 2020-07-08
> @ursabot benchmark --suite-filter=arrow-compute-aggregate-benchmark

The result of the buildbot, https://ci.ursalabs.org/#/builders/73/builds/93. The clang compiler can generate more better SIMD instructions.

Below is the results for the two typical null proportions(0% 0.01%).
```
31       SumKernelInt32/1048576/0    6.576 GiB/sec   19.756 GiB/sec   200.444  {'run_name': 'SumKernelInt32/1048576/0', 'run_...
14       SumKernelInt16/1048576/0    3.314 GiB/sec    8.518 GiB/sec   157.026  {'run_name': 'SumKernelInt16/1048576/0', 'run_...
15       SumKernelFloat/1048576/0    8.647 GiB/sec   20.204 GiB/sec   133.656  {'run_name': 'SumKernelFloat/1048576/0', 'run_...
34   SumKernelFloat/1048576/10000    6.669 GiB/sec   14.262 GiB/sec   113.867  {'run_name': 'SumKernelFloat/1048576/10000', '...
24        SumKernelInt8/1048576/0    1.906 GiB/sec    3.794 GiB/sec    99.079  {'run_name': 'SumKernelInt8/1048576/0', 'run_t...
23       SumKernelInt64/1048576/0   18.130 GiB/sec   25.094 GiB/sec    38.410  {'run_name': 'SumKernelInt64/1048576/0', 'run_...
29      SumKernelDouble/1048576/0   18.296 GiB/sec   24.188 GiB/sec    32.204  {'run_name': 'SumKernelDouble/1048576/0', 'run...
19   SumKernelInt32/1048576/10000   10.897 GiB/sec   14.163 GiB/sec    29.967  {'run_name': 'SumKernelInt32/1048576/10000', '...
4    SumKernelInt64/1048576/10000   16.745 GiB/sec   20.665 GiB/sec    23.411  {'run_name': 'SumKernelInt64/1048576/10000', '...
26  SumKernelDouble/1048576/10000   16.439 GiB/sec   19.635 GiB/sec    19.441  {'run_name': 'SumKernelDouble/1048576/10000', ...
9    SumKernelInt16/1048576/10000    5.870 GiB/sec    6.403 GiB/sec     9.069  {'run_name': 'SumKernelInt16/1048576/10000', '...
11    SumKernelInt8/1048576/10000    2.958 GiB/sec    2.917 GiB/sec    -1.382  {'run_name': 'SumKernelInt8/1048576/10000', 'r...
```

For other 1% 50% null case, there are some small regression(20%) which are expected as BitBlockCounter used now.