@wesm 2020-07-08
I'm tied up until mid-day today but plan to review this afternoon (US time)