@pitrou 2020-07-08
The approach looks ok to me. I'll take a closer look later.