@pitrou 2020-07-08
I see similar benchmark results on an AMD Ryzen (AVX2).